Started work on 0.1.1 with a bug fix

+++ main.js
@@ -504,8 +504,10 @@ ipcMain.on('uninstall-game', (event, gameId, gameState, gameTitle) => {
 });
 
 function forceStopDownload() {
-    downloadProcess.kill();
-    downloadProcess = null;
+    if (downloadProcess) {
+        downloadProcess.kill();
+        downloadProcess = null;
+    }
     if (inDownload) {
         if (extractionActive) {
             if (readStream) readStream.destroy();
